refactor: styles 폴더 생성, 흡연자 비흡연자 알림 설정 스타일 컴포넌트 분리
