+++ src/pages/MyPage/SetNotifyNonSmoker.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
+import { SelectContainer, SharedBox, DotsBox, BtnBox, Btn, SelectBtn } from '../../styles/NotifyStyle';
 import SetNotifyHeader from "./SetNotifyHeader";
 import { putMyPageDistAPI } from "../../apis/api";
-import dotsImg from '../../assets/background_dots.png';
 
 const SetNotifyNonSmoker = () => {
   const navigate = useNavigate();
@@ -73,94 +73,8 @@ const SetNotifyNonSmoker = () => {
 
 export default SetNotifyNonSmoker;
 
-const SelectContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-
-  img {
-    width: 10%;
-  }
-`;
-
-const Box = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  width: 90%;
-  height: 15rem;
-  margin: 5% 0 15% 0;
-  border-radius: 0.5rem;
-  background-color: #FFF100;
-
+const Box = styled(SharedBox)`
   h3 {
     font-size: 1.2rem;
   }
-`;
-
-const DotsBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-image: url(${dotsImg});  
-  background-size: 50%;  
-  background-repeat: no-repeat; 
-  background-position: left top; 
-  text-align: center;
-  width: 96%;
-  height: 94%;
-`;
-
-const BtnBox = styled.div`
-  display: flex;
-  justify-content: space-evenly;
-  width: 90%;
-  margin-top: 2rem;
-  margin-bottom: 1rem;
-`;
-
-const Btn = styled.div`
-  border: 2px solid transparent;
-  border-radius: 0.4rem;
-  padding: 0.8rem 1.2rem;
-  color: white;
-  background-color: #272A30;
-  font-weight: bold;
-  
-  cursor: pointer;
-
-  &:hover,
-  &:focus {
-    color: #272A30;
-    background-color: #f7f152;
-
-    outline: none;
-    border: 2px solid #272A30;
-  }
-
-  &.active {
-    color: #272A30;
-    background-color: #f7f152;
-    border: 2px solid #272A30;
-  }
-`;
-
-const SelectBtn = styled.div`
-  width: 90%;
-  border: 2px solid black;
-  border-radius: 0.3rem;
-  padding: 1rem 0;
-  text-align: center;
-  font-weight: bold;
-  color: white;
-  background-color: #272A30;
-  box-shadow: 0.2rem 0.2rem 0.2rem #FEFBBD;
-
-  cursor: pointer;
 `;
