+++ src/pages/MyPage/SetType.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import SharedHeader from "../../components/SharedHeader";
-import { Container } from "../../components/SharedContainer";
+import { Container } from "../../styles/SharedContainer";
 import smokerImg from "../../assets/smoker.png";
 import nonSmokerImg from "../../assets/nonSmoker.png";
 import detailBackgroundImage from '../../assets/mypage_detail_background.png';
