+++ src/styles/SharedContainer.js
